compiled for array of numbers

+++ alx-low_level_programming/0x0F-function_pointers/1-array_iterator.c
@@ -1,8 +1,9 @@
+#include <stdlib.h>
 #include "function_pointers.h"
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
 	for(i = 0; i < size; i++)
 	{
